renamed proj & finished expl int analyzer
