+++ Robust.Shared/Analyzers/RequiresExplicitImplementationAttribute.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Robust.Shared
+namespace Robust.Shared.Analyzers
 {
     [AttributeUsage(AttributeTargets.Interface)]
     public class RequiresExplicitImplementationAttribute : Attribute { }
